rename guards to utils
