+++ .changeset/shy-buses-mix.md
@@ -2,4 +2,4 @@
 '@backstage/plugin-scaffolder-react': minor
 ---
 
-refactor `createAsyncValidators` to be recursive
+refactor `createAsyncValidators` to be recursive to ensure validators are called in nested schemas.
