+++ plugins/scaffolder-react/src/next/components/Stepper/utils.ts
